Put nameserver info on all subinterfaces of a vif
At the moment nova-agent only puts the nameserver info on one of the
subinterfaces of a vif seemingly picked at random. Alot of times
this ends up being the ipv6 configuration section, which causes
the nameserver information to not be loaded when IPv6 is disabled
on the virtual machine. This commit puts the nameserver information
on all of the subinterfaces to prevent this issue from occurring.